requirements.txt
PyQt6>=6.4.0
mutagen>=1.46.0
requests>=2.28.0
Pillow>=10.0.0
qtawesome>=1.3.0
sounddevice>=0.4.6
soundfile>=0.12.1
